+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### v 21 (WIP)
+
+- new: snackbar, when download completed
+
 ### v 20 Bertha von Suttner (Die Waffen nieder)
 
 - new: disable webview diagnostic data